resources/requirements.txt
pywhatkit
speechrecognition
pyaudio
openai
azure-cognitiveservices-speech
